[Refactor] Request와 Response 관련 Enum 클래스들 Enums 디렉토리로 이동
프로젝트의 가독성과 구조적 명확성을 향상시키기 위해, HttpRequest와 HttpResponse와 관련된 Enum 클래스들을 별도의 'enums' 디렉토리로 이동했습니다.
