+++ src/main/java/webserver/http/request/enums/HttpMethod.java
@@ -1,4 +1,4 @@
-package webserver.http.request;
+package webserver.http.request.enums;
 
 public enum HttpMethod {
     GET, POST, PUT, DELETE, PATCH
